requirements.txt
annotated-types==0.6.0
anyio==4.3.0
ariadne==0.22
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
strawberry-graphql==0.220.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
